requirements.txt
streamlit
ultralytics
Pillow
opencv-python-headless>=4.5.0
numpy


pafy
youtube_dl

